Combining `Description` and `DescriptionItem`.
